Make platform/store.ts a façade over the two store implementations
store.ts was 2812 lines: a 335-line re-export prelude and the Store aggregate
interface, then class InMemoryStore and class FirestoreStore. The two classes
move to store/in-memory.ts and store/firestore.ts and are re-exported, so all
229 importers stay untouched.

Both bodies moved unchanged — comment- and blank-stripped, they are identical
to their previous text, 954 and 1010 code lines. Each file's imports were
derived from store.ts's own import map and confirmed by tsc, not by eye.

store.ts is now 302 lines, and its two ratchet baselines reseal down with it
rather than leaving room to regrow.
