+++ eslint-rules/comment-prose-baseline.json
@@ -277,7 +277,7 @@
     "apps/api/src/platform/spa-paths.test.ts": 101,
     "apps/api/src/platform/spa-paths.ts": 437,
     "apps/api/src/platform/store.test.ts": 394,
-    "apps/api/src/platform/store.ts": 817,
+    "apps/api/src/platform/store.ts": 625,
     "apps/api/src/platform/submission-status.test.ts": 35,
     "apps/api/src/platform/submission-status.ts": 257,
     "apps/api/src/platform/tar.test.ts": 104,
@@ -313,6 +313,8 @@
     "apps/api/src/store-firestore.test.ts": 566,
     "apps/api/src/store/fake-firestore.ts": 328,
     "apps/api/src/store/firestore-util.ts": 163,
+    "apps/api/src/store/firestore.ts": 145,
+    "apps/api/src/store/in-memory.ts": 40,
     "apps/api/src/store/records/access-tokens.ts": 204,
     "apps/api/src/store/records/agent-keys.ts": 169,
     "apps/api/src/store/records/build-log.ts": 492,
