+++ eslint-rules/module-size-baseline.json
@@ -397,7 +397,7 @@
     "apps/api/src/platform/spa-paths.test.ts": 119,
     "apps/api/src/platform/spa-paths.ts": 135,
     "apps/api/src/platform/store.test.ts": 799,
-    "apps/api/src/platform/store.ts": 2812,
+    "apps/api/src/platform/store.ts": 302,
     "apps/api/src/platform/submission-status.test.ts": 138,
     "apps/api/src/platform/submission-status.ts": 234,
     "apps/api/src/platform/submission-token.test.ts": 24,
@@ -436,6 +436,8 @@
     "apps/api/src/self-build.test.ts": 1655,
     "apps/api/src/static-serving.test.ts": 95,
     "apps/api/src/store-firestore.test.ts": 710,
+    "apps/api/src/store/firestore.ts": 1323,
+    "apps/api/src/store/in-memory.ts": 1259,
     "apps/api/src/submissions.test.ts": 7368,
     "apps/api/src/submissions.ts": 1693,
     "apps/api/src/tab-complete-deploy-flag.test.ts": 28,
